src/test/java/com/psh10066/refactoring/_16_temporary_field/_36_introduce_special_case/CustomerServiceTest.java
package com.psh10066.refactoring._16_temporary_field._36_introduce_special_case;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertTrue;

class CustomerServiceTest {

    CustomerService customerService = new CustomerService();
    Customer unknown = new Customer("unknown", null, null);

    BillingPlan specialBillingPlan = new BillingPlan();
    Customer special = new Customer("special", specialBillingPlan, new PaymentHistory(1));

    @Test
    void customerName() {
        String unknownCustomerName = customerService.customerName(new Site(unknown));
        assertEquals("occupant", unknownCustomerName);

        String customer = customerService.customerName(new Site(special));
        assertEquals("special", customer);
    }

    @Test
    void billingPlan() {
        assertTrue((customerService.billingPlan(new Site(unknown)) instanceof BasicBillingPlan));

        assertEquals(specialBillingPlan, customerService.billingPlan(new Site(special)));
    }

    @Test
    void weeksDelingquent() {
        assertEquals(1, customerService.weeksDelinquent(new Site(special)));

        assertEquals(0, customerService.weeksDelinquent(new Site(unknown)));
    }
}